move stuff into `gltf_util`
